backend/requirements.txt
alembic==1.11.1
anyio==3.7.1
APScheduler==3.10.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==3.0.1
asttokens==2.2.1
attrs==23.1.0
autoflake==2.2.1
autopep8==2.0.4
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
bidict==0.22.1
black==23.10.1
bleach==6.0.0
blinker==1.6.2
cachelib==0.10.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
comm==0.1.3
commonmark==0.9.1
cryptography==42.0.7
debugpy==1.6.7
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
dill==0.3.7
dnspython==2.6.1
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.17.1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-Session==0.5.0
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.5
fqdn==1.5.1
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
Mako==1.2.4
Markdown==3.5
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.0.1
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.1
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.25.1
openpyxl==3.1.2
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
paramiko==3.4.0
parso==0.8.3
passlib==1.7.4
pathspec==0.11.2
pdoc3==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.1
pluggy==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
Pygments==2.15.1
PyJWT==2.7.0
pylint==3.0.2
pymongo==4.8.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-json-logger==2.0.7
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.29.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rpds-py==0.8.10
Send2Trash==1.8.2
six==1.16.0
skimpy==0.0.8
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tomlkit==0.12.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typeguard==2.13.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
unicodecsv==0.14.1
uri-template==1.3.0
urllib3==2.0.5
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
widgetsnbextension==4.0.8
xlrd==2.0.1
